pyproject.toml
[project]
name = "ebsagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncio>=4.0.0",
    "fastmcp>=2.12.2",
    "oci>=2.160.0",
]
